move raw cmds to own file
